- renamed `SlickSugar` to `SlickUtil` - moved `paramPlaceholder` to `SlickUtil`
