+++ app/src/main/scala/org/alephium/explorer/persistence/queries/TokenQueries.scala
@@ -26,7 +26,7 @@ import org.alephium.explorer.api.model._
 import org.alephium.explorer.persistence._
 import org.alephium.explorer.persistence.schema.CustomGetResult._
 import org.alephium.explorer.persistence.schema.CustomSetParameter._
-import org.alephium.explorer.util.SlickSugar._
+import org.alephium.explorer.util.SlickUtil._
 import org.alephium.util.{TimeStamp, U256}
 
 object TokenQueries extends StrictLogging {
