+++ app/src/main/scala/org/alephium/explorer/service/TransactionHistoryService.scala
@@ -35,7 +35,7 @@ import org.alephium.explorer.persistence.schema.CustomGetResult._
 import org.alephium.explorer.persistence.schema.CustomJdbcTypes._
 import org.alephium.explorer.persistence.schema.CustomSetParameter._
 import org.alephium.explorer.util.Scheduler
-import org.alephium.explorer.util.SlickSugar._
+import org.alephium.explorer.util.SlickUtil._
 import org.alephium.explorer.util.TimeUtil._
 import org.alephium.protocol.ALPH
 import org.alephium.util.{Duration, TimeStamp}
